packages/core: move out showLoginPopup from OAuthRequestApi to separate lib module + tweak and document payload format
